Immense refactoring
wvlt namespace
deprecating old matrices
splines
int -> uint
deleted: makepa, cfg, eyes
split net.cpp into some sources
new names to library.h sources

+++ myLib_dataHandler.cpp
@@ -65,7 +65,7 @@ void writePlainData(const QString outPath,
     }
     else
     {
-        numOfSlices = min(numOfSlices,
+        numOfSlices = min(uint(numOfSlices),
                           data.cols() - start);
     }
